refactor: :hammer: Improved the code structure by seperating the modules
